Updates based on Pull Request Comments Part 3
