+++ monitors/c/docs/dev/engines.md
@@ -3,10 +3,12 @@
 Each engine is identified by a constant value:
 
 ```C
-const uint8_t R2U2_ENG_CG: u8 = 2; // Immediate Configuration Directive
+// static const uint8_t R2U2_ENG_NA = 0; // Null instruction tag - acts as ENDSEQ - not utilized since v4.0
+// static const uint8_t R2U2_ENG_SY = 1; // System commands - reserved for monitor control - not utilized since v4.0
+static const uint8_t R2U2_ENG_CG = 2; // Immediate Configuration Directive
 // Original Atomic Checker was 3, but has been removed since v4.0
-const uint8_t R2U2_ENG_TL: u8 = 4; // MLTL Temporal logic engine
-const uint8_t R2U2_ENG_BZ: u8 = 5; // Booleanizer
+static const uint8_t  R2U2_ENG_TL = 4; // MLTL Temporal logic engine
+static const uint8_t R2U2_ENG_BZ = 5; // Booleanizer
 ```
 
 Of these three values, only two are "real" (BZ and TL), while CG is used internally by the monitor for signaling.
