+++ monitors/c/src/engines/engines.h
@@ -4,8 +4,8 @@
 #include "../internals/errors.h"
 #include "../memory/monitor.h"
 
-// static const uint8_t R2U2_ENG_NA = 0; // Null instruction tag - acts as ENDSEQ
-// static const uint8_t R2U2_ENG_SY = 1; // System commands - reserved for monitor control
+// static const uint8_t R2U2_ENG_NA = 0; // Null instruction tag - acts as ENDSEQ - not utilized since v4.0
+// static const uint8_t R2U2_ENG_SY = 1; // System commands - reserved for monitor control - not utilized since v4.0
 static const uint8_t R2U2_ENG_CG = 2; // Immediate Configuration Directive
 // Original Atomic Checker was 3, but has been removed since v4.0
 static const uint8_t  R2U2_ENG_TL = 4; // MLTL Temporal logic engine
