+++ monitors/rust/r2u2_core/src/engines/mod.rs
@@ -15,8 +15,8 @@ use cortex_m_semihosting::hprintln;
 #[cfg(feature = "debug_print_std")]
 use libc_print::std_name::println;
 
-// pub const R2U2_ENG_NA: u8 = 0; // Null instruction tag - acts as ENDSEQ
-// pub const R2U2_ENG_SY: u8 = 1; // System commands - reserved for monitor control
+// pub const R2U2_ENG_NA: u8 = 0; // Null instruction tag - acts as ENDSEQ - not utilized since v4.0
+// pub const R2U2_ENG_SY: u8 = 1; // System commands - reserved for monitor control - not utilized since v4.0
 pub const R2U2_ENG_CG: u8 = 2; // Immediate Configuration Directive
 // Original Atomic Checker was 3, but has been removed since v4.0
 pub const R2U2_ENG_TL: u8 = 4; // MLTL Temporal logic engine
